Move helpers for partioned data to coreutils
