+++ packages/op-coreutils/src/op_coreutils/partitioned/__init__.py
@@ -1,6 +1,9 @@
 """Utilities for partitioned storage."""
 
 from .breakout import breakout_partitions
-from .marker import Marker
+from .location import DataLocation, MarkersLocation
+from .marker import Marker, markers_for_dates, marker_exists
 from .output import KeyValue, WrittenParquetPath
+from .status import all_outputs_complete
 from .types import SinkMarkerPath, SinkOutputRootPath
+from .writer import write_single_part, write_marker
