+++ packages/op-datasets/src/op_datasets/etl/ingestion/status.py
@@ -2,47 +2,12 @@ from datetime import timedelta
 
 from op_coreutils import clickhouse
 from op_coreutils.logger import structlog
-from op_coreutils.partitioned import SinkMarkerPath
 from op_coreutils.time import datetime_fromepoch, now_seconds
 
 from .batches import BlockBatch
-from .sinks import RawOnchainDataSink
-from .utilities import RawOnchainDataLocation, RawOnchainDataProvider
+from .sources import RawOnchainDataProvider
 
 log = structlog.get_logger()
-
-
-def all_outputs_complete(
-    sinks: list[RawOnchainDataLocation], markers: list[SinkMarkerPath]
-) -> bool:
-    """Check if all outputs are complete.
-
-    This function is somewhat low-level in that it receives the explicit completion
-    markers that we are looking for. It checks that those markers are present in all
-    of the data sinks.
-    """
-    result = True
-    for data_location in sinks:
-        sink = RawOnchainDataSink(location=data_location)
-        complete = []
-        incomplete = []
-        for marker in markers:
-            if sink.is_complete(marker):
-                complete.append(marker)
-            else:
-                incomplete.append(marker)
-
-        log.info(
-            f"{len(complete)} complete, {len(incomplete)} incomplete on sink={sink.location.name}"
-        )
-
-        if incomplete:
-            log.info(f"Showing the first 5 incomplete locations at {sink.location.name}")
-            for marker_location in sorted(incomplete)[:5]:
-                log.info(f"DataSink {sink.location.name!r} is incomplete at {marker_location!r}")
-            result = False
-
-    return result
 
 
 def all_inputs_ready(provider: RawOnchainDataProvider, block_batch: BlockBatch) -> bool:
